Nerfing the power of Lonsdaleite and Perfect Lonsdaleite Tools.
This is done to keep the progression in line with other mods that add armor and tools, such as Better End, Biomes You'll Go, etc. Further consideration was made based on the difficulty to acquire the various items. The massive jump between in power between Lonsdaleite and Perfect Lonsdaleite was not warranted by simply requiring additional smelting.